make sha regex search somewhat format-agnostic
By simplifying junegunn's regex search, we arrive at a simple hash regex
that looks for a hash-like string for 7 characters or more.
That way, no matter what the git log format is, as long as the commit is
mention on the line, we should be able to find it.

A nice addition to that would be to make the format a global setting to
be change on the user's vimrc, and to make the syntax file more
accepting of different formats.