Improve handling of cmdline-special escape in GT command
There were some issue when trying to run commands like:
:GT --follow -- %
Where we would hope that the log tree tracking the current file would
display, but the escaping of the '%' cmdline-special would not execute
properly, since the `expand()` command expect a string to start with
such a character.
Also added some better handling if the file itself contains a "%", which
I often encounter (e.g. in yocto)